pyproject.toml
[project]
name = "poker-planning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.112.3",
    "uvicorn>=0.30.6",
    "jinja2>=3.1.4",
    "pydantic>=2.8.2",
    "python-multipart>=0.0.9",
    "rich>=13.8.0",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.2",
    "pymercure>=0.0.1.0",
    "pyjwt>=2.9.0",
]
